lesson_29/requirements.txt
psycopg2-binary==2.9.9
allure-pytest==2.14.2
pytest==8.3.4